BBS만들기 - 2023.10.05 (삭제, 수정 기능 추가)

+++ BBS/src/main/webapp/view.jsp
@@ -125,7 +125,7 @@
 					if(userID != null && userID.equals(bbs.getUserID())) {
 				%>
 						<a href="update.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">수정</a>
-						<a href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
+						<a onClick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
 				<%
 					}
 				%>
